split public and gated APIs into their own files for clarity.
